Implemented a big part of the renderer! Pulsing USRGames Color. Welcome to Entropy v1.0

+++ src/Entropy/Core/Math/Quaternion.cpp
@@ -4,7 +4,7 @@
 
 float Quaternion::length()
 {
-    return (float)sqrt(X * X + Y * Y + Z * Z + W * W);
+    return (float)sqrtf(X * X + Y * Y + Z * Z + W * W);
 }
 
 Quaternion& Quaternion::normalize()
